Adding Update Product Logic without Unit Test (Total 80% Processed)

+++ src/main/java/com/senior/candleShopProject/datasource/domain/products/IProductResp.java
@@ -7,7 +7,7 @@ public interface IProductResp {
     UUID getProductId();
     String getProductName();
     BigDecimal getPrice();
-    Integer getWeight();
+    double getWeight();
     String getDescription();
     String getSlug();
     String getIsActive();
